==> code/0_Load_Packages.R <==
library(RColorBrewer)
library(DT)
library(dplyr)

# library(readxl)
# library(RSQLite)
# library(RpostGIS) # https://github.com/solymosin/RpostGIS

#library(phyloseq)
# library(genefilter)
#library(microbiome)
# library(tidyr)
#library(tidyverse)
# library(ape)
#library(vegan)
#library(ggnewscale)
#library(DESeq2)
#library(xtable)
#library(DT)

#library(ggplot2)
#library(readxl)
#library(RSQLite)
#library(rgdal)
#library(tmap)
#library(tmaptools)
#library(raster)
#library(sf)
#library(fields)
#library(spdep)
#library(SpatialEpi)
#library(sp)
# library(RNetCDF)
#library(weathermetrics)
#library(R2WinBUGS)
#library(spsurvey)
